pyproject.toml
[project]
name = "email-triage-env"
version = "1.0.0"
description = "OpenEnv benchmark environment for realistic email triage"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openenv-core[core]>=0.2.0",
  "fastapi>=0.110.0",
  "uvicorn>=0.29.0",
  "pydantic>=2.7.0",
  "requests>=2.31.0",
  "openai>=1.30.0",
]

[project.scripts]
server = "server.app:main"
